Big Ol' Calendar Fix
Yours truly was a bit stupid when building V1 and thought 'there's only one view and no need to change the date, a static HTML calendar will be easier' but it turns out that it's still loads easier to make all the various features of fullcalendar work when you let it generate the calendar rather than just adding the FC classes to the HTML yourself... who'd've thought, ha. On the surface it looks very much the same but it should be a lot easier to implement FC features now.