repairs and enhancements to src edit modal
Addresses RCAL-10 issue with regard to Validating sources.

There is essentially no work to do now for validation since
the source is correct upon construction: you cannot save a
source that is invalid using the source modal dialog.
Key to this was fixing the artist/album/track fields to a
nominal value when the source is known to be a directory/file.
Moreover, the name of a source may not be edited after it has
been saved, nor can certain defining features (like type).
We might just remove the validate button, but I have not yet
done so.

Also fixed a blunder in initializing play lists selector.

Selecting an album will update the encoding in the dialog.

Finally, added a button in the modal header that allows a
source to be deleted. The required changes to the data model
have NOT been implemented yet.  The delete button does not
appear if the source has not been registered yet (new sources).