support day repeat checkboxes
This addresses RCAL-14 and RCAL-21.

Checking or unchecking day-of-week checkboxes in the event dialog
will create/destroy peer events, where peers are events that agree
in title, start, and end times. Newly cloned events are actually
independent and can be adjusted separately.

Likely this is not quite perfect--we need to test the edge cases.