auto resize events of known duration
As soon as a source with known duration is dragged into
a schedule, its end point is adjusted to reflect its duration.
After compacting, there will be no gaps if the finite
source is dropped onto a stream like a radio station.
This covers the most common use case in rsked:
create the radio streams then drop albums and playlists
over them.

Also, the title of the calendar includes "Modified" if
the schedule has been changed since load. Not all possible
avenues of change are currently being trapped, so beware.